refactor(tab): export tab insteadof tabs pane
