+++ package.json
@@ -2,7 +2,7 @@
   "name": "taro-material",
   "private": true,
   "author": "lico",
-  "version": "0.6.0-10",
+  "version": "0.6.0-11",
   "description": "Mini Program components that implement Google's Material Design.",
   "keywords": [
     "react",
