+++ packages/taro-material/src/index.js
@@ -101,7 +101,7 @@ export { default as RMWatermark } from './Watermark';
 export { default as RMNoticeBar } from './NoticeBar';
 export { default as RMNotice } from './Notice';
 export { default as RMTabs } from './Tabs';
-export { default as RMTabsPane } from './TabsPane';
+export { default as RMTab } from './Tab';
 export { default as RMMeter } from './Meter';
 export { default as RMStarVote } from './StarVote';
 export { default as RMSteps } from './Steps';
