+++ pages/pages/material/tabs/index.js
@@ -1,11 +1,10 @@
 import Taro, { Component } from '@tarojs/taro';
-import { View, Button, Canvas } from '@tarojs/components';
+import { View } from '@tarojs/components';
 
-import { RMTabs, RMTabsPane } from '../../../';
+import { RMTabs, RMTab } from '../../../';
 
 import RMPage from '../../../Page';
 
-import theme from '../../../styles/theme';
 import './index.scss';
 
 const allNavList = [
@@ -124,14 +123,9 @@ class Index extends RMPage {
               centered
             >
               {navList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">default - no scroll - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -147,14 +141,9 @@ class Index extends RMPage {
               color="default"
             >
               {navList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">default - no scroll - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -170,14 +159,9 @@ class Index extends RMPage {
               color="primary"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">primary - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -193,14 +177,9 @@ class Index extends RMPage {
               color="secondary"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">secondary - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -216,14 +195,9 @@ class Index extends RMPage {
               color="inherit"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">inherit - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -239,14 +213,9 @@ class Index extends RMPage {
               color="action"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">action - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -262,14 +231,9 @@ class Index extends RMPage {
               color="error"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">error - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -285,14 +249,9 @@ class Index extends RMPage {
               color="disabled"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">disabled - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -308,14 +267,9 @@ class Index extends RMPage {
               color="success"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">success - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -331,14 +285,9 @@ class Index extends RMPage {
               color="warning"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">warning - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -354,14 +303,9 @@ class Index extends RMPage {
               color="progress"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane
-                  tabDirection="vertical"
-                  current={current}
-                  index={index}
-                  key={item.title}
-                >
+                <RMTab tabDirection="vertical" current={current} index={index} key={item.title}>
                   <View style="height:200px;">progress - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -380,9 +324,9 @@ class Index extends RMPage {
               centered
             >
               {navList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">default - no scroll - centered - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -397,9 +341,9 @@ class Index extends RMPage {
               centered={false}
             >
               {navList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">default - no scroll - not centered - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -413,9 +357,9 @@ class Index extends RMPage {
               color="primary"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">primary - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -429,9 +373,9 @@ class Index extends RMPage {
               color="secondary"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">secondary - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -445,9 +389,9 @@ class Index extends RMPage {
               color="inherit"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">inherit - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -461,9 +405,9 @@ class Index extends RMPage {
               color="action"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">action - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -477,9 +421,9 @@ class Index extends RMPage {
               color="error"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">error - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -493,9 +437,9 @@ class Index extends RMPage {
               color="disabled"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">disabled - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -509,9 +453,9 @@ class Index extends RMPage {
               color="success"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">success - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -525,9 +469,9 @@ class Index extends RMPage {
               color="warning"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">warning - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
@@ -541,9 +485,9 @@ class Index extends RMPage {
               color="progress"
             >
               {allNavList.map((item, index) => (
-                <RMTabsPane current={current} index={index} key={item.title}>
+                <RMTab current={current} index={index} key={item.title}>
                   <View style="height:200px;">progress - {current}</View>
-                </RMTabsPane>
+                </RMTab>
               ))}
             </RMTabs>
           </View>
